feat, refactor #31: RetrievalAdapter properly implemented
- This piece is needed for computing Rank@K metric
- Testing is still needed
